requirements.txt
attrs==21.4.0
Authlib==1.2.0
aws-cdk-lib==2.17.0
aws-cdk.aws-lambda-python-alpha==2.17.0a0
awslambdaric==2.0.1
boto3==1.21.8
botocore==1.24.8
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
constructs==10.0.54
cryptography==39.0.2
cycler==0.11.0
exceptiongroup==1.1.0
fonttools==4.29.1
idna==3.3
iniconfig==2.0.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsii==1.55.1
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.1.0
multitasking==0.0.10
packaging==21.3
pluggy==1.0.0
publication==0.0.3
pycoingecko==2.2.0
pycparser==2.21
pyparsing==3.0.7
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
simplejson==3.17.2
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
